moo-1.2
Changes:
- Change internal storage from int to int64_t. (Thanks mbalmer@!)
- Prevent division by zeroes. (Thanks again mbalmer@!)

+++ extern.h
@@ -1,9 +1,9 @@
-/*	$Id: extern.h,v 1.2 2006/03/04 04:39:13 ray Exp $	*/
+/*	$Id: extern.h,v 1.3 2006/06/16 07:32:27 ray Exp $	*/
 
 /*
  * Written by Raymond Lai <ray@cyth.net>.
  * Public domain.
  */
 
-int getbin(const char *);
-int getnum(const char *);
+int64_t getbin(const char *);
+int64_t getnum(const char *);
